belt-block: move tests to /tests folder
